dashboard/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.0
psycopg2-binary>=2.9.9
pytz>=2024.1
